Move custom map library generation tied to InstrumentationContext to allow for re-use; Add on-demand type specialization for custom map setters when certain array literals are passed in
