+++ src/instrumenter/custom_maps.ts
@@ -1,746 +1,37 @@
 import {
-    AddressType,
-    ArrayType,
     ArrayTypeName,
     assert,
     Assignment,
     ASTNodeFactory,
-    Block,
-    BoolType,
-    BytesType,
     ContractDefinition,
-    ContractKind,
-    DataLocation,
     eq,
     Expression,
     ExpressionStatement,
     FunctionCall,
     FunctionCallKind,
     FunctionDefinition,
-    FunctionKind,
-    FunctionStateMutability,
-    FunctionVisibility,
     Identifier,
     IndexAccess,
-    IntType,
-    LiteralKind,
     Mapping,
-    MappingType,
     MemberAccess,
-    Mutability,
-    PointerType,
     replaceNode,
     SourceUnit,
-    Statement,
-    StateVariableVisibility,
-    StringType,
     StructDefinition,
     TypeName,
     typeNameToTypeNode,
-    TypeNode,
     UnaryOperation,
-    UserDefinedType,
     UserDefinedTypeName,
-    VariableDeclaration,
-    VariableDeclarationStatement,
-    variableDeclarationToTypeNode
+    VariableDeclaration
 } from "solc-typed-ast";
 import {
     ConcreteDatastructurePath,
     explodeTupleAssignment,
     findStateVarUpdates,
     single,
-    transpileType
+    StateVarRefDesc
 } from "..";
 import { InstrumentationContext } from "./instrumentation_context";
 import { InstrumentationSiteType } from "./transpiling_context";
-
-export type StateVarRefDesc = [Expression, VariableDeclaration, ConcreteDatastructurePath];
-
-function getTypeDesc(typ: TypeNode): string {
-    if (
-        typ instanceof IntType ||
-        typ instanceof StringType ||
-        typ instanceof BytesType ||
-        typ instanceof BoolType
-    ) {
-        return typ.pp();
-    }
-
-    if (typ instanceof AddressType) {
-        return "address" + (typ.payable ? "_payable" : "");
-    }
-
-    if (typ instanceof ArrayType) {
-        return `${getTypeDesc(typ.elementT)}_arr` + (typ.size !== undefined ? `_${typ.size}` : "");
-    }
-
-    if (typ instanceof UserDefinedType) {
-        return `${typ.name.replace(".", "_")}_${typ.definition.id}`;
-    }
-
-    if (typ instanceof MappingType) {
-        return `mapping_${getTypeDesc(typ.keyType)}_to_${getTypeDesc(typ.valueType)}`;
-    }
-
-    if (typ instanceof PointerType) {
-        return getTypeDesc(typ.to);
-    }
-
-    throw new Error(`Unknown type ${typ.pp()} in getTypeDesc`);
-}
-
-function getLibraryName(keyT: TypeNode, valueT: TypeNode): string {
-    return `${getTypeDesc(keyT)}_to_${getTypeDesc(valueT)}`;
-}
-
-function addStructField(
-    factory: ASTNodeFactory,
-    name: string,
-    typ: TypeNode,
-    struct: StructDefinition
-): void {
-    const field = factory.makeVariableDeclaration(
-        false,
-        false,
-        name,
-        struct.id,
-        false,
-        DataLocation.Default,
-        StateVariableVisibility.Default,
-        Mutability.Mutable,
-        "<missing>",
-        undefined,
-        transpileType(typ, factory)
-    );
-    struct.appendChild(field);
-}
-
-function needsLocation(t: TypeNode): boolean {
-    if (
-        t instanceof ArrayType ||
-        t instanceof StringType ||
-        t instanceof BytesType ||
-        t instanceof MappingType ||
-        (t instanceof UserDefinedType && t.definition instanceof StructDefinition)
-    ) {
-        return true;
-    }
-
-    return false;
-}
-
-function addFunArg(
-    factory: ASTNodeFactory,
-    name: string,
-    typ: TypeNode,
-    location: DataLocation,
-    fun: FunctionDefinition
-): VariableDeclaration {
-    const arg = factory.makeVariableDeclaration(
-        false,
-        false,
-        name,
-        fun.id,
-        false,
-        location,
-        StateVariableVisibility.Default,
-        Mutability.Mutable,
-        "<missing>",
-        undefined,
-        transpileType(typ, factory)
-    );
-
-    fun.vParameters.appendChild(arg);
-    return arg;
-}
-
-function addFunRet(
-    factory: ASTNodeFactory,
-    name: string,
-    typ: TypeNode,
-    location: DataLocation,
-    fun: FunctionDefinition
-): VariableDeclaration {
-    const arg = factory.makeVariableDeclaration(
-        false,
-        false,
-        name,
-        fun.id,
-        false,
-        location,
-        StateVariableVisibility.Default,
-        Mutability.Mutable,
-        "<missing>",
-        undefined,
-        transpileType(typ, factory)
-    );
-
-    fun.vReturnParameters.appendChild(arg);
-    return arg;
-}
-
-function makeStruct(
-    factory: ASTNodeFactory,
-    keyT: TypeNode,
-    valueT: TypeNode,
-    lib: ContractDefinition
-): StructDefinition {
-    const struct = factory.makeStructDefinition("S", "S", lib.id, "", []);
-
-    addStructField(factory, "innerM", new MappingType(keyT, valueT), struct);
-    addStructField(factory, "keys", new ArrayType(keyT), struct);
-    addStructField(factory, "keyIdxM", new MappingType(keyT, new IntType(256, false)), struct);
-
-    return struct;
-}
-
-function addEmptyFun(
-    factory: ASTNodeFactory,
-    name: string,
-    container: ContractDefinition | SourceUnit
-): FunctionDefinition {
-    const fun = factory.makeFunctionDefinition(
-        container.id,
-        FunctionKind.Function,
-        name,
-        false,
-        FunctionVisibility.Public,
-        FunctionStateMutability.NonPayable,
-        false,
-        factory.makeParameterList([]),
-        factory.makeParameterList([]),
-        [],
-        undefined,
-        factory.makeBlock([])
-    );
-    container.appendChild(fun);
-    return fun;
-}
-
-function addStmt(
-    factory: ASTNodeFactory,
-    fun: FunctionDefinition,
-    arg: Statement | Expression
-): void {
-    const body = fun.vBody as Block;
-    body.appendChild(arg instanceof Statement ? arg : factory.makeExpressionStatement(arg));
-}
-
-function mkStructFieldAcc(
-    factory: ASTNodeFactory,
-    base: Expression,
-    struct: StructDefinition,
-    idx: number
-): MemberAccess {
-    const field = struct.vMembers[idx];
-    return factory.makeMemberAccess("<missing>", base, field.name, field.id);
-}
-
-function mkInnerM(
-    factory: ASTNodeFactory,
-    base: Expression,
-    struct: StructDefinition
-): MemberAccess {
-    return mkStructFieldAcc(factory, base, struct, 0);
-}
-
-function mkVarDecl(
-    factory: ASTNodeFactory,
-    name: string,
-    typ: TypeName,
-    location: DataLocation,
-    val: Expression,
-    fun: FunctionDefinition
-): [VariableDeclaration, VariableDeclarationStatement] {
-    const decl = factory.makeVariableDeclaration(
-        false,
-        false,
-        name,
-        fun.id,
-        false,
-        location,
-        StateVariableVisibility.Default,
-        Mutability.Mutable,
-        "<missing>",
-        undefined,
-        typ
-    );
-
-    const stmt = factory.makeVariableDeclarationStatement([decl.id], [decl], val);
-
-    return [decl, stmt];
-}
-
-function makeIncDecFun(
-    factory: ASTNodeFactory,
-    keyT: TypeNode,
-    valueT: TypeNode,
-    struct: StructDefinition,
-    lib: ContractDefinition,
-    operator: "++" | "--",
-    prefix: boolean
-): FunctionDefinition {
-    const name = (operator == "++" ? "inc" : "dec") + (prefix ? "_pre" : "");
-    const fun = addEmptyFun(factory, name, lib);
-
-    const m = addFunArg(
-        factory,
-        "m",
-        new UserDefinedType(struct.name, struct),
-        DataLocation.Storage,
-        fun
-    );
-    const key = addFunArg(
-        factory,
-        "key",
-        keyT,
-        needsLocation(keyT) ? DataLocation.Memory : DataLocation.Default,
-        fun
-    );
-
-    const RET = addFunRet(
-        factory,
-        "RET",
-        valueT,
-        needsLocation(valueT) ? DataLocation.Storage : DataLocation.Default,
-        fun
-    );
-
-    const mkInnerM = () => mkStructFieldAcc(factory, factory.makeIdentifierFor(m), struct, 0);
-    const curVal = factory.makeIndexAccess("<missing>", mkInnerM(), factory.makeIdentifierFor(key));
-    const newVal = factory.makeBinaryOperation(
-        "<missing>",
-        operator[0],
-        curVal,
-        factory.makeLiteral("<missing>", LiteralKind.Number, "", "1")
-    );
-    const setter = single(lib.vFunctions.filter((f) => f.name == "set"));
-
-    const update = factory.makeFunctionCall(
-        "<missing>",
-        FunctionCallKind.FunctionCall,
-        factory.makeIdentifierFor(setter),
-        [factory.makeIdentifierFor(m), factory.makeIdentifierFor(key), newVal]
-    );
-
-    if (prefix) {
-        addStmt(factory, fun, factory.makeReturn(fun.vReturnParameters.id, update));
-    } else {
-        addStmt(
-            factory,
-            fun,
-            factory.makeAssignment(
-                "<missing>",
-                "=",
-                factory.makeIdentifierFor(RET),
-                factory.makeIndexAccess("<missing>", mkInnerM(), factory.makeIdentifierFor(key))
-            )
-        );
-        addStmt(factory, fun, update);
-    }
-
-    return fun;
-}
-
-function getLoc(t: TypeNode, defLoc: DataLocation): DataLocation {
-    if (t instanceof PointerType) {
-        return t.location;
-    }
-
-    return needsLocation(t) ? defLoc : DataLocation.Default;
-}
-
-function makeGetFun(
-    factory: ASTNodeFactory,
-    keyT: TypeNode,
-    valueT: TypeNode,
-    struct: StructDefinition,
-    lib: ContractDefinition
-): FunctionDefinition {
-    const fun = addEmptyFun(factory, "get", lib);
-
-    const m = addFunArg(
-        factory,
-        "m",
-        new UserDefinedType(struct.name, struct),
-        DataLocation.Storage,
-        fun
-    );
-    const key = addFunArg(factory, "key", keyT, getLoc(keyT, DataLocation.Memory), fun);
-    addFunRet(factory, "", valueT, getLoc(valueT, DataLocation.Storage), fun);
-
-    // return m.innerM[key];
-    addStmt(
-        factory,
-        fun,
-        factory.makeReturn(
-            fun.vReturnParameters.id,
-            factory.makeIndexAccess(
-                "<missing>",
-                mkInnerM(factory, factory.makeIdentifierFor(m), struct),
-                factory.makeIdentifierFor(key)
-            )
-        )
-    );
-
-    return fun;
-}
-
-function makeSetFun(
-    factory: ASTNodeFactory,
-    keyT: TypeNode,
-    valueT: TypeNode,
-    struct: StructDefinition,
-    lib: ContractDefinition
-): FunctionDefinition {
-    const fun = addEmptyFun(factory, "set", lib);
-
-    const m = addFunArg(
-        factory,
-        "m",
-        new UserDefinedType(struct.name, struct),
-        DataLocation.Storage,
-        fun
-    );
-    const key = addFunArg(factory, "key", keyT, getLoc(keyT, DataLocation.Memory), fun);
-    const val = addFunArg(factory, "val", valueT, getLoc(valueT, DataLocation.Memory), fun);
-
-    addFunRet(factory, "", valueT, getLoc(valueT, DataLocation.Storage), fun);
-
-    const mkInnerM = () => mkStructFieldAcc(factory, factory.makeIdentifierFor(m), struct, 0);
-    const mkKeys = () => mkStructFieldAcc(factory, factory.makeIdentifierFor(m), struct, 1);
-    const mkKeysLen = () => factory.makeMemberAccess("<missing>", mkKeys(), "length", -1);
-    const mkKeysPush = () => factory.makeMemberAccess("<missing>", mkKeys(), "push", -1);
-    const mkKeyIdxM = () => mkStructFieldAcc(factory, factory.makeIdentifierFor(m), struct, 2);
-
-    //m.innerM[key] = val;
-    addStmt(
-        factory,
-        fun,
-        factory.makeAssignment(
-            "<missing>",
-            "=",
-            factory.makeIndexAccess("<missing>", mkInnerM(), factory.makeIdentifierFor(key)),
-            factory.makeIdentifierFor(val)
-        )
-    );
-
-    //uint idx = m.keyIdxM[key];
-    const idx = factory.makeVariableDeclaration(
-        false,
-        false,
-        "idx",
-        fun.id,
-        false,
-        DataLocation.Default,
-        StateVariableVisibility.Default,
-        Mutability.Mutable,
-        "<missing>",
-        undefined,
-        factory.makeElementaryTypeName("<missing>", "uint256")
-    );
-
-    addStmt(
-        factory,
-        fun,
-        factory.makeVariableDeclarationStatement(
-            [idx.id],
-            [idx],
-            factory.makeIndexAccess("<missing>", mkKeyIdxM(), factory.makeIdentifierFor(key))
-        )
-    );
-
-    //if (idx > 0) {
-    //    return;
-    //}
-    addStmt(
-        factory,
-        fun,
-        factory.makeIfStatement(
-            factory.makeBinaryOperation(
-                "<missing>",
-                ">",
-                factory.makeIdentifierFor(idx),
-                factory.makeLiteral("<missing>", LiteralKind.Number, "", "0")
-            ),
-            factory.makeReturn(
-                fun.vReturnParameters.id,
-                factory.makeIndexAccess("<missing>", mkInnerM(), factory.makeIdentifierFor(key))
-            )
-        )
-    );
-
-    //if (m.keys.length == 0) {
-    //    m.keys.push();
-    //}
-    addStmt(
-        factory,
-        fun,
-        factory.makeIfStatement(
-            factory.makeBinaryOperation(
-                "<missing>",
-                "==",
-                mkKeysLen(),
-                factory.makeLiteral("<missing>", LiteralKind.Number, "", "0")
-            ),
-            factory.makeBlock([
-                factory.makeExpressionStatement(
-                    factory.makeFunctionCall(
-                        "<missing>",
-                        FunctionCallKind.FunctionCall,
-                        mkKeysPush(),
-                        []
-                    )
-                )
-            ])
-        )
-    );
-
-    //m.keyIdxM[key] = m.keys.length;
-    addStmt(
-        factory,
-        fun,
-        factory.makeAssignment(
-            "<missing>",
-            "=",
-            factory.makeIndexAccess("<missing>", mkKeyIdxM(), factory.makeIdentifierFor(key)),
-            mkKeysLen()
-        )
-    );
-    //m.keys.push(key);
-    addStmt(
-        factory,
-        fun,
-        factory.makeExpressionStatement(
-            factory.makeFunctionCall("<missing>", FunctionCallKind.FunctionCall, mkKeysPush(), [
-                factory.makeIdentifierFor(key)
-            ])
-        )
-    );
-
-    // return m.innerM[key];
-    addStmt(
-        factory,
-        fun,
-        factory.makeReturn(
-            fun.vReturnParameters.id,
-            factory.makeIndexAccess("<missing>", mkInnerM(), factory.makeIdentifierFor(key))
-        )
-    );
-
-    return fun;
-}
-
-function makeDeleteFun(
-    factory: ASTNodeFactory,
-    keyT: TypeNode,
-    valueT: TypeNode,
-    struct: StructDefinition,
-    lib: ContractDefinition
-): FunctionDefinition {
-    const fun = addEmptyFun(factory, "deleteKey", lib);
-
-    const m = addFunArg(
-        factory,
-        "m",
-        new UserDefinedType(struct.name, struct),
-        DataLocation.Storage,
-        fun
-    );
-    const key = addFunArg(factory, "key", keyT, getLoc(keyT, DataLocation.Memory), fun);
-
-    const mkInnerM = () => mkStructFieldAcc(factory, factory.makeIdentifierFor(m), struct, 0);
-    const mkKeys = () => mkStructFieldAcc(factory, factory.makeIdentifierFor(m), struct, 1);
-    const mkKeysLen = () => factory.makeMemberAccess("<missing>", mkKeys(), "length", -1);
-    const mkKeysPop = () => factory.makeMemberAccess("<missing>", mkKeys(), "pop", -1);
-    const mkKeyIdxM = () => mkStructFieldAcc(factory, factory.makeIdentifierFor(m), struct, 2);
-    const mkDelete = (exp: Expression) =>
-        factory.makeExpressionStatement(
-            factory.makeUnaryOperation("<missing>", true, "delete", exp)
-        );
-
-    // delete m.innerM[key];
-    addStmt(
-        factory,
-        fun,
-        mkDelete(factory.makeIndexAccess("<missing>", mkInnerM(), factory.makeIdentifierFor(key)))
-    );
-
-    // uint idx = m.keyIdxM[key];
-    const [idx, declStmt] = mkVarDecl(
-        factory,
-        "idx",
-        factory.makeElementaryTypeName("<missing>", "uint256"),
-        DataLocation.Default,
-        factory.makeIndexAccess("<missing>", mkKeyIdxM(), factory.makeIdentifierFor(key)),
-        fun
-    );
-    addStmt(factory, fun, declStmt);
-
-    // if (idx == 0) {
-    //     return;
-    // }
-    addStmt(
-        factory,
-        fun,
-        factory.makeIfStatement(
-            factory.makeBinaryOperation(
-                "<missing>",
-                "==",
-                factory.makeIdentifierFor(idx),
-                factory.makeLiteral("<missing>", LiteralKind.Number, "", "0")
-            ),
-            factory.makeReturn(fun.vReturnParameters.id)
-        )
-    );
-
-    // if (idx != m.keys.length - 1) {
-    const cond = factory.makeBinaryOperation(
-        "<missing>",
-        "!=",
-        factory.makeIdentifierFor(idx),
-        factory.makeBinaryOperation(
-            "<missing>",
-            "-",
-            mkKeysLen(),
-            factory.makeLiteral("<missing>", LiteralKind.Number, "", "1")
-        )
-    );
-
-    const ifBody: Statement[] = [];
-    //     uint lastKey = m.keys[m.keys.length - 1];
-    const [lastKey, lastKeyDecl] = mkVarDecl(
-        factory,
-        "lastKey",
-        transpileType(keyT, factory),
-        getLoc(keyT, DataLocation.Storage),
-        factory.makeIndexAccess(
-            "<missing>",
-            mkKeys(),
-            factory.makeBinaryOperation(
-                "<missing>",
-                "-",
-                mkKeysLen(),
-                factory.makeLiteral("<missing>", LiteralKind.Number, "", "1")
-            )
-        ),
-        fun
-    );
-    ifBody.push(lastKeyDecl);
-
-    //     m.keys[idx] = lastKey;
-    ifBody.push(
-        factory.makeExpressionStatement(
-            factory.makeAssignment(
-                "<missing>",
-                "=",
-                factory.makeIndexAccess("<missing>", mkKeys(), factory.makeIdentifierFor(idx)),
-                factory.makeIdentifierFor(lastKey)
-            )
-        )
-    );
-
-    //     m.keyIdxM[lastKey] = idx;
-    ifBody.push(
-        factory.makeExpressionStatement(
-            factory.makeAssignment(
-                "<missing>",
-                "=",
-                factory.makeIndexAccess(
-                    "<missing>",
-                    mkKeyIdxM(),
-                    factory.makeIdentifierFor(lastKey)
-                ),
-                factory.makeIdentifierFor(idx)
-            )
-        )
-    );
-
-    // }
-    addStmt(factory, fun, factory.makeIfStatement(cond, factory.makeBlock(ifBody)));
-
-    // m.keys.pop();
-    addStmt(
-        factory,
-        fun,
-        factory.makeFunctionCall("<missing>", FunctionCallKind.FunctionCall, mkKeysPop(), [])
-    );
-    // delete m.keyIdxM[key];
-    addStmt(
-        factory,
-        fun,
-        mkDelete(factory.makeIndexAccess("<mising>", mkKeyIdxM(), factory.makeIdentifierFor(key)))
-    );
-
-    return fun;
-}
-
-function typeContainsMap(t: TypeNode, compilerVersion: string): boolean {
-    if (t instanceof MappingType) {
-        return true;
-    }
-
-    if (t instanceof ArrayType) {
-        return typeContainsMap(t.elementT, compilerVersion);
-    }
-
-    if (t instanceof UserDefinedType && t.definition instanceof StructDefinition) {
-        for (const field of t.definition.vMembers) {
-            const fieldT = variableDeclarationToTypeNode(field);
-            if (typeContainsMap(fieldT, compilerVersion)) {
-                return true;
-            }
-        }
-    }
-
-    if (t instanceof PointerType) {
-        return typeContainsMap(t.to, compilerVersion);
-    }
-
-    return false;
-}
-
-export function generateMapLibrary(
-    factory: ASTNodeFactory,
-    keyT: TypeNode,
-    valueT: TypeNode,
-    container: SourceUnit,
-    compilerVersion: string
-): ContractDefinition {
-    const libName = getLibraryName(keyT, valueT);
-
-    const lib = factory.makeContractDefinition(
-        libName,
-        container.id,
-        ContractKind.Library,
-        false,
-        true,
-        [],
-        []
-    );
-    container.appendChild(lib);
-
-    const struct = makeStruct(factory, keyT, valueT, lib);
-    lib.appendChild(struct);
-    makeGetFun(factory, keyT, valueT, struct, lib);
-
-    // For value types containing maps its not possible to re-assign or delete indices
-    if (!typeContainsMap(valueT, compilerVersion)) {
-        makeSetFun(factory, keyT, valueT, struct, lib);
-        makeDeleteFun(factory, keyT, valueT, struct, lib);
-    }
-
-    // For numeric types emit helpers for ++,--, {+,-,*,/,**,%,<<,>>}=
-    if (valueT instanceof IntType) {
-        makeIncDecFun(factory, keyT, valueT, struct, lib, "++", true);
-        makeIncDecFun(factory, keyT, valueT, struct, lib, "++", false);
-        makeIncDecFun(factory, keyT, valueT, struct, lib, "--", true);
-        makeIncDecFun(factory, keyT, valueT, struct, lib, "--", false);
-    }
-
-    return lib;
-}
 
 type DatastructurePath = Array<null | string>;
 
@@ -829,49 +120,28 @@ function splitExpr(e: Expression): [Expression, Expression] {
     return [e.vBaseExpression, e.vIndexExpression];
 }
 
-function getDeleteKey(factory: ASTNodeFactory, lib: ContractDefinition): Expression {
-    const deleteKeyF = single(lib.vFunctions.filter((fun) => fun.name === "deleteKey"));
+function mkFunRef(factory: ASTNodeFactory, fn: FunctionDefinition): MemberAccess {
     return factory.makeMemberAccess(
         "<missing>",
-        factory.makeIdentifierFor(lib),
-        "deleteKey",
-        deleteKeyF.id
+        factory.makeIdentifierFor(fn.vScope as ContractDefinition),
+        fn.name,
+        fn.id
     );
 }
 
-function getSetter(factory: ASTNodeFactory, lib: ContractDefinition): Expression {
-    const setter = single(lib.vFunctions.filter((fun) => fun.name === "set"));
-    return factory.makeMemberAccess("<missing>", factory.makeIdentifierFor(lib), "set", setter.id);
-}
-
-function getIncDecFun(
-    factory: ASTNodeFactory,
-    lib: ContractDefinition,
-    operator: "++" | "--",
-    prefix: boolean
-): Expression {
-    const name = (operator == "++" ? "inc" : "dec") + (prefix ? "_pre" : "");
-    const setter = single(lib.vFunctions.filter((fun) => fun.name === name));
-    return factory.makeMemberAccess("<missing>", factory.makeIdentifierFor(lib), name, setter.id);
-}
-
-function getGetter(factory: ASTNodeFactory, lib: ContractDefinition): Expression {
-    const getter = single(lib.vFunctions.filter((fun) => fun.name === "get"));
-    return factory.makeMemberAccess("<missing>", factory.makeIdentifierFor(lib), "get", getter.id);
-}
-
 function replaceAssignmentHelper(
-    factory: ASTNodeFactory,
+    instrCtx: InstrumentationContext,
     assignment: Assignment,
     lib: ContractDefinition
 ): void {
+    const factory = instrCtx.factory;
     const newVal = assignment.vRightHandSide;
     const [base, index] = splitExpr(assignment.vLeftHandSide);
 
     const newNode = factory.makeFunctionCall(
         "<missing>",
         FunctionCallKind.FunctionCall,
-        getSetter(factory, lib),
+        mkFunRef(factory, instrCtx.getCustomMapSetter(lib, newVal)),
         [base, index, newVal]
     );
 
@@ -970,7 +240,6 @@ export function findStateVarReferences(
 export function interposeMap(
     instrCtx: InstrumentationContext,
     targets: Array<[VariableDeclaration, DatastructurePath]>,
-    mapContainer: SourceUnit,
     units: SourceUnit[]
 ): void {
     const allUpdates = findStateVarUpdates(units);
@@ -981,7 +250,6 @@ export function interposeMap(
         lookupPathInType(stateVar.vType as TypeName, path)
     );
     const factory = instrCtx.factory;
-    const compilerVersion = instrCtx.compilerVersion;
 
     for (let i = 0; i < targets.length; i++) {
         const stateVar = targets[i][0];
@@ -997,8 +265,8 @@ export function interposeMap(
         const valueT = typeNameToTypeNode(mapT.vValueType);
 
         // 0. Generate custom library implementation
-        const lib = generateMapLibrary(factory, keyT, valueT, mapContainer, compilerVersion);
-        const struct = single(lib.vStructs);
+        const lib = instrCtx.getCustomMapLibrary(keyT, valueT);
+        const struct = instrCtx.getCustomMapStruct(lib);
 
         // 1. Replace the type of `T` in the `stateVar` declaration with `L.S`
         const newMapT = factory.makeUserDefinedTypeName(
@@ -1025,7 +293,7 @@ export function interposeMap(
 
                 // Simple non-tuple case
                 if (lhsPath.length === 0) {
-                    replaceAssignmentHelper(factory, assignment, lib);
+                    replaceAssignmentHelper(instrCtx, assignment, lib);
                 } else {
                     // Tuple assignment case.
                     // @todo Do we need a new instrumentation type here?
@@ -1040,7 +308,7 @@ export function interposeMap(
                         assignment
                     )) {
                         if (eq(tuplePath, lhsPath)) {
-                            replaceAssignmentHelper(factory, tempAssignment, lib);
+                            replaceAssignmentHelper(instrCtx, tempAssignment, lib);
                         }
                     }
                 }
@@ -1049,7 +317,7 @@ export function interposeMap(
 
                 if (updateNode.operator === "delete") {
                     assert(updateNode.parent instanceof ExpressionStatement, ``);
-                    const deleteKeyF = getDeleteKey(factory, lib);
+                    const deleteKeyF = mkFunRef(factory, instrCtx.getCustomMapDeleteKey(lib));
 
                     const newNode = factory.makeFunctionCall(
                         "<missing>",
@@ -1060,12 +328,11 @@ export function interposeMap(
                     replaceNode(updateNode, newNode);
                 } else {
                     assert(updateNode.operator === "++" || updateNode.operator == "--", ``);
-                    const incDecF = getIncDecFun(
+                    const incDecF = mkFunRef(
                         factory,
-                        lib,
-                        updateNode.operator,
-                        updateNode.prefix
+                        instrCtx.getCustomMapIncDec(lib, updateNode.operator, updateNode.prefix)
                     );
+
                     const newNode = factory.makeFunctionCall(
                         "<missing>",
                         FunctionCallKind.FunctionCall,
@@ -1094,7 +361,7 @@ export function interposeMap(
         // replace the occuring references with calls to `L.get()`
         for (const [refNode] of findStateVarReferences(units, stateVar, path)) {
             const [base, index] = splitExpr(refNode);
-            const getterF = getGetter(factory, lib);
+            const getterF = mkFunRef(factory, instrCtx.getCustomMapGetter(lib));
             const newNode = factory.makeFunctionCall(
                 "<misisng>",
                 FunctionCallKind.FunctionCall,
