+++ src/instrumenter/index.ts
@@ -4,4 +4,5 @@ export * from "./instrument";
 export * from "./transpile";
 export * from "./state_vars";
 export * from "./state_var_instrumenter";
+export * from "./custom_maps_templates";
 export * from "./custom_maps";
